+++ client/helpers/router.js
@@ -140,6 +140,7 @@ var filters = {
 
   canView: function() {
     if(Session.get('settingsLoaded') && !canView()){
+      console.log('cannot view')
       this.render('no_rights');
       this.stop();
     }
@@ -174,7 +175,6 @@ var filters = {
   hasCompletedProfile: function() {
     var user = Meteor.user();
     if (user && ! Meteor.loggingIn() && ! userProfileComplete(user)){
-      // Session.set('selectedUserId', user._id);
       this.render('user_email');
       this.stop();
     }
@@ -186,7 +186,14 @@ var filters = {
 
 Router.load( function () {
   clearSeenErrors(); // set all errors who have already been seen to not show anymore
-  Session.set('categorySlug', null); 
+  Session.set('categorySlug', null);
+
+  // if we're not on the search page itself, clear search query and field
+  if(getCurrentRoute().indexOf('search') == -1){
+    Session.set('searchQuery', '');
+    $('.search-field').val('').blur();
+  }
+
 });
 
 // Before Hooks
@@ -198,7 +205,8 @@ Router.before(filters.nProgressHook, {only: [
   'posts_best', 
   'posts_pending', 
   'posts_digest',
-  'posts_category', 
+  'posts_category',
+  'search',
   'post_page',
   'post_edit',
   'comment_page',
@@ -209,9 +217,9 @@ Router.before(filters.nProgressHook, {only: [
   'all-users'
 ]});
 
+Router.before(filters.canView);
 Router.before(filters.hasCompletedProfile);
 Router.before(filters.isLoggedIn, {only: ['comment_reply','post_submit']});
-Router.before(filters.canView);
 Router.before(filters.isLoggedOut, {only: ['signin', 'signup']});
 Router.before(filters.canPost, {only: ['posts_pending', 'comment_reply', 'post_submit']});
 Router.before(filters.canEditPost, {only: ['post_edit']});
@@ -240,30 +248,24 @@ PostsListController = RouteController.extend({
   template:'posts_list',
   waitOn: function () {
     // take the first segment of the path to get the view, unless it's '/' in which case the view default to 'top'
-    var view = this.path == '/' ? 'top' : this.path.split('/')[1];
-    var limit = this.params.limit || getSetting('postsPerPage', 10);
     // note: most of the time this.params.slug will be empty
-    var parameters = getParameters(view, limit, this.params.slug);
+    this._terms = {
+      view: this.path == '/' ? 'top' : this.path.split('/')[1],
+      limit: this.params.limit || getSetting('postsPerPage', 10),
+      category: this.params.slug,
+      query: Session.get("searchQuery")
+    }
     return [
-      Meteor.subscribe('postsList', parameters.find, parameters.options),
-      Meteor.subscribe('postsListUsers', parameters.find, parameters.options)
+      Meteor.subscribe('postsList', this._terms),
+      Meteor.subscribe('postsListUsers', this._terms)
     ]
   },
   data: function () {
-    var view = this.path == '/' ? 'top' : this.path.split('/')[1],
-        limit = this.params.limit || getSetting('postsPerPage', 10),
-        parameters = getParameters(view, limit, this.params.slug),
+    var parameters = getParameters(this._terms),
         posts = Posts.find(parameters.find, parameters.options);
         postsCount = posts.count();
-        
-    Session.set('postsLimit', limit);
-
-    // get posts and decorate them with rank property
-    // note: not actually used; 
-    // posts = posts.map(function (post, index) {
-    //   post.rank = index;
-    //   return post;
-    // });
+  
+    Session.set('postsLimit', this._terms.limit);
 
     return {
       postsList: posts,
@@ -282,16 +284,25 @@ PostsDigestController = RouteController.extend({
   template: 'posts_digest',
   waitOn: function() {
     // if day is set, use that. If not default to today
-    var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today');
-    var parameters = getDigestParameters(currentDate);
+    var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today'),
+        terms = {
+          view: 'digest',
+          after: moment(currentDate).startOf('day').valueOf(),
+          before: moment(currentDate).endOf('day').valueOf()
+        }
     return [
-      Meteor.subscribe('postsList', parameters.find, parameters.options),
-      Meteor.subscribe('postsListUsers', parameters.find, parameters.options)
+      Meteor.subscribe('postsList', terms),
+      Meteor.subscribe('postsListUsers', terms)
     ]
   },
   data: function() {
-    var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today');
-    var parameters = getDigestParameters(currentDate);
+    var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today'),
+        terms = {
+          view: 'digest',
+          after: moment(currentDate).startOf('day').valueOf(),
+          before: moment(currentDate).endOf('day').valueOf()
+        },
+        parameters = getParameters(terms);
     Session.set('currentDate', currentDate);
     return {
       posts: Posts.find(parameters.find, parameters.options)
@@ -349,8 +360,13 @@ UserPageController = RouteController.extend({
   data: function() {
     var findById = Meteor.users.findOne(this.params._idOrSlug);
     var findBySlug = Meteor.users.findOne({slug: this.params._idOrSlug});
-    return {
-      user: (typeof findById == "undefined") ? findBySlug : findById
+    if(typeof findById !== "undefined"){
+      // redirect to slug-based URL
+      Router.go(getProfileUrl(findById), {replaceState: true});
+    }else{
+      return {
+        user: (typeof findById == "undefined") ? findBySlug : findById
+      }
     }
   }
 });
@@ -360,6 +376,8 @@ UserPageController = RouteController.extend({
 //--------------------------------------------------------------------------------------------------//
 
 Router.map(function() {
+
+  // -------------------------------------------- Post Lists -------------------------------------------- //
 
   // Top
 
@@ -406,6 +424,12 @@ Router.map(function() {
 
   // TODO: enable /category/new, /category/best, etc. views
 
+  // Search
+
+  this.route('search', {
+    path: '/search/:limit?',
+    controller: PostsListController    
+  });
 
   // Digest
 
@@ -520,9 +544,8 @@ Router.map(function() {
     path: '/all-users/:limit?',
     template: 'users',
     waitOn: function() {
-      var limit = parseInt(this.params.limit) || 20,
-          parameters = getUsersParameters(this.params.filterBy, this.params.sortBy, limit);
-      return Meteor.subscribe('allUsers', parameters.find, parameters.options);
+      var limit = parseInt(this.params.limit) || 20;
+      return Meteor.subscribe('allUsers', this.params.filterBy, this.params.sortBy, limit);
     },
     data: function() {
       var limit = parseInt(this.params.limit) || 20,