+++ client/views/users/user_edit.js
@@ -17,14 +17,17 @@ Template.user_edit.helpers({
   profileUrl: function(){
     return Meteor.absoluteUrl()+"users/"+this.slug;
   },
+  hasNotificationsUsers : function(){
+    return getUserSetting('notifications.users', '', this) ? 'checked' : '';
+  },
   hasNotificationsPosts : function(){
-    return getUserSetting('notifications.posts') ? 'checked' : '';
+    return getUserSetting('notifications.notifications_posts', '', this) ? 'checked' : '';
   },
   hasNotificationsComments : function(){
-    return getUserSetting('notifications.comments') ? 'checked' : '';
+    return getUserSetting('notifications.comments', '', this) ? 'checked' : '';
   },
   hasNotificationsReplies : function(){
-    return getUserSetting('notifications.replies') ? 'checked' : '';
+    return getUserSetting('notifications.replies', '', this) ? 'checked' : '';
   }
 })
 
@@ -45,10 +48,11 @@ Template.user_edit.events = {
       "profile.twitter": $target.find('[name=twitter]').val(),
       "profile.github": $target.find('[name=github]').val(),
       "profile.site": $target.find('[name=site]').val(),
+      "profile.notifications.users": $('input[name=notifications_users]:checked').length, // only actually used for admins
       "profile.notifications.posts": $('input[name=notifications_posts]:checked').length,
       "profile.notifications.comments": $('input[name=notifications_comments]:checked').length,
       "profile.notifications.replies": $('input[name=notifications_replies]:checked').length,
-      "invitesCount": parseInt($target.find('[name=invitesCount]').val())
+      "inviteCount": parseInt($target.find('[name=inviteCount]').val())
     };
 
     var old_password = $target.find('[name=old_password]').val();