+++ client/views/users/user_item.js
@@ -20,35 +20,22 @@ Template.user_item.helpers({
   userIsAdmin: function(){
     return isAdmin(this);
   },
-  profileUrl: function () {
+  getProfileUrl: function () {
     return getProfileUrl(this);
   },
   getKarma: function() {
     return Math.round(100*this.karma)/100;
+  },
+  getInvitedUserProfileUrl: function () {
+    var user = Meteor.users.findOne(this.invitedId);
+    return getProfileUrl(user);
   }
 });
 
 Template.user_item.events({
   'click .invite-link': function(e, instance){
     e.preventDefault();
-    var user = Meteor.users.findOne(instance.data._id);
-    Meteor.users.update(user._id,{
-      $set:{
-        isInvited: true
-      }
-    }, {multi: false}, function(error){
-      if(error){
-        throwError();
-      }else{
-        Meteor.call('createNotification', {
-          event: 'accountApproved', 
-          properties: {}, 
-          userToNotify: user, 
-          userDoingAction: Meteor.user(), 
-          sendEmail: getSetting("emailNotifications")
-        });
-      }
-    });
+    Meteor.call('inviteUser', instance.data._id);
   },
   'click .uninvite-link': function(e, instance){
     e.preventDefault();