+++ collections/notifications.js
@@ -9,44 +9,71 @@ Notifications.allow({
   , remove: canEditById
 });
 
-getNotification = function(event, properties, context){
-  var notification = {};
-  // the default context to display notifications is the notification sidebar
-  var context = typeof context === 'undefined' ? 'sidebar' : context;
-  var p = properties;
+getNotificationContents = function(notification, context){
+  // the same notifications can be displayed in multiple contexts: on-site in the sidebar, sent by email, etc. 
+  var event = notification.event,
+      p = notification.properties,
+      context = typeof context === 'undefined' ? 'sidebar' : context,
+      userToNotify = Meteor.users.findOne(notification.userId);
+
+
   switch(event){
     case 'newReply':
-      notification.subject = i18n.t('Someone replied to your comment on')+' "'+p.postHeadline+'"';
-      notification.text = p.commentAuthorName+i18n.t(' has replied to your comment on')+' "'+p.postHeadline+'": '+getPostCommentUrl(p.postId, p.commentId);
-      notification.html = '<p><a href="'+getUserUrl(p.commentAuthorId)+'">'+p.commentAuthorName+'</a>'+i18n.t(' has replied to your comment on')+' "<a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">'+p.postHeadline+'</a>"</p>';
-      if(context === 'email')
-        notification.html += '<p>'+p.commentExcerpt+'</p><a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">'+i18n.t('Read more')+'</a>';
-    break;
+      var n = {
+        subject: 'Someone replied to your comment on "'+p.postHeadline+'"',
+        text: p.commentAuthorName+' has replied to your comment on "'+p.postHeadline+'": '+getPostCommentUrl(p.postId, p.commentId),
+        html: '<p><a href="'+getProfileUrlById(p.commentAuthorId)+'">'+p.commentAuthorName+'</a> has replied to your comment on "<a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">'+p.postHeadline+'</a>"</p>'
+      }
+      if(context == 'email')
+        n.html += '<p>'+p.commentExcerpt+'</p><a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">Read more</a>';
+      break;
 
     case 'newComment':
-      notification.subject = i18n.t('A new comment on your post')+' "'+p.postHeadline+'"';
-      notification.text = i18n.t('You have a new comment by ')+p.commentAuthorName+i18n.t(' on your post')+' "'+p.postHeadline+'": '+getPostCommentUrl(p.postId, p.commentId);
-      notification.html = '<p><a href="'+getUserUrl(p.commentAuthorId)+'">'+p.commentAuthorName+'</a> left a new comment on your post "<a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">'+p.postHeadline+'</a>"</p>';
-      if(context === 'email')
-        notification.html += '<p>'+p.commentExcerpt+'</p><a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">'+i18n.t('Read more')+'</a>';
-    break;
+      var n = {
+        subject: 'A new comment on your post "'+p.postHeadline+'"',
+        text: 'You have a new comment by '+p.commentAuthorName+' on your post "'+p.postHeadline+'": '+getPostCommentUrl(p.postId, p.commentId),
+        html: '<p><a href="'+getProfileUrlById(p.commentAuthorId)+'">'+p.commentAuthorName+'</a> left a new comment on your post "<a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">'+p.postHeadline+'</a>"</p>'
+      }
+      if(context == 'email')
+        n.html += '<p>'+p.commentExcerpt+'</p><a href="'+getPostCommentUrl(p.postId, p.commentId)+'" class="action-link">Read more</a>';
+      break;
 
     case 'newPost':
-      notification.subject = p.postAuthorName+i18n.t(' has created a new post')+': "'+p.postHeadline+'"';
-      notification.text = p.postAuthorName+i18n.t(' has created a new post')+': "'+p.postHeadline+'" '+getPostUrl(p.postId);
-      notification.html = '<a href="'+getUserUrl(p.postAuthorId)+'">'+p.postAuthorName+'</a>'+i18n.t(' has created a new post')+': "<a href="'+getPostUrl(p.postId)+'" class="action-link">'+p.postHeadline+'</a>".';
-    break;
+      var n = {
+        subject: p.postAuthorName+' has created a new post: "'+p.postHeadline+'"',
+        text: p.postAuthorName+' has created a new post: "'+p.postHeadline+'" '+getPostUrl(p.postId),
+        html: '<a href="'+getProfileUrlById(p.postAuthorId)+'">'+p.postAuthorName+'</a> has created a new post: "<a href="'+getPostUrl(p.postId)+'" class="action-link">'+p.postHeadline+'</a>".'  
+      }
+      break;
 
     case 'accountApproved':
-      notification.subject = i18n.t('Your account has been approved.');
-      notification.text = i18n.t('Welcome to ')+getSetting('title')+'! '+i18n.t('Your account has just been approved.');
-      notification.html = i18n.t('Welcome to ')+getSetting('title')+'!<br/> '+i18n.t('Your account has just been approved.')+' <a href="'+Meteor.absoluteUrl()+'">'+i18n.t('Start posting.')+'</a>';
-    break;
+      var n = {
+        subject: 'Your account has been approved.',
+        text: 'Welcome to '+getSetting('title')+'! Your account has just been approved.',
+        html: 'Welcome to '+getSetting('title')+'!<br/> Your account has just been approved. <a href="'+Meteor.absoluteUrl()+'">Start posting.</a>'
+      }
+      break;
+
+    case 'newUser':
+      var n = {
+        subject: 'New user: '+p.username,
+        text: 'A new user account has been created: '+p.profileUrl,
+        html: 'A new user account has been created: <a href="'+p.profileUrl+'">'+p.username+'</a>'
+      }
+      break;
 
     default:
     break;
   }
-  return notification;
+
+  // if context is email, append unsubscribe link to all outgoing notifications
+  if(context == 'email'){
+    n.to = getEmail(userToNotify);
+    n.text = n.text + '\n\n Unsubscribe from all notifications: '+getUnsubscribeLink(userToNotify);
+    n.html = n.html + '<br/><br/><a href="'+getUnsubscribeLink(userToNotify)+'">Unsubscribe from all notifications</a>';
+  }
+
+  return n;
 }
 
 Meteor.methods({