+++ server/publications.js
@@ -3,7 +3,7 @@ var privacyOptions = { // false means private
   isAdmin: false,
   emails: false,
   notifications: false,
-  invitesCount: false,
+  inviteCount: false,
   'profile.email': false,
   'services.twitter.accessToken': false,
   'services.twitter.accessTokenSecret': false,
@@ -31,6 +31,7 @@ Meteor.publish('singleUser', function(userIdOrSlug) {
     // if we find something when treating the argument as an ID, return that; else assume it's a slug
     return findById.count() ? findById : findBySlug;
   }
+  return [];
 });
 
 // Publish authors of the current post and its comments
@@ -51,6 +52,7 @@ Meteor.publish('postUsers', function(postId) {
     
     return Meteor.users.find({_id: {$in: users}}, {fields: privacyOptions});
   }
+  return [];
 });
 
 // Publish author of the current comment
@@ -60,34 +62,42 @@ Meteor.publish('commentUser', function(commentId) {
     var comment = Comments.findOne(commentId);
     return Meteor.users.find({_id: comment && comment.userId}, {fields: privacyOptions});
   }
+  return [];
 });
 
 // Publish all the users that have posted the currently displayed list of posts
 
-Meteor.publish('postsListUsers', function(find, options) {
+Meteor.publish('postsListUsers', function(terms) {
   if(canViewById(this.userId)){
-    var posts = Posts.find(find, options);
-    var userIds = _.pluck(posts.fetch(), 'userId');
+    var parameters = getParameters(terms),
+        posts = Posts.find(parameters.find, parameters.options),
+        userIds = _.pluck(posts.fetch(), 'userId');
     return Meteor.users.find({_id: {$in: userIds}}, {fields: privacyOptions, multi: true});
   }
+  return [];
 });
 
 // Publish all users
 
-Meteor.publish('allUsers', function(find, options) {
+Meteor.publish('allUsers', function(filterBy, sortBy, limit) {
   if(canViewById(this.userId)){
+    var parameters = getUsersParameters(filterBy, sortBy, limit);
     if (!isAdminById(this.userId)) // if user is not admin, filter out sensitive info
-      options = _.extend(options, {fields: privacyOptions});
-    return Meteor.users.find(find, options);  
+      parameters.options = _.extend(parameters.options, {fields: privacyOptions});
+    return Meteor.users.find(parameters.find, parameters.options);  
   }
+  return [];
 });
 
 // publish all users for admins to make autocomplete work
 // TODO: find a better way
 
 Meteor.publish('allUsersAdmin', function() {
-  if (isAdminById(this.userId))
-    return Meteor.users.find();  
+  if (isAdminById(this.userId)) {
+    return Meteor.users.find();
+  } else {
+    return [];
+  }
 });
 
 // -------------------------------------------- Posts -------------------------------------------- //
@@ -98,6 +108,7 @@ Meteor.publish('singlePost', function(id) {
   if(canViewById(this.userId)){
     return Posts.find(id);
   }
+  return [];
 });
 
 // Publish the post related to the current comment
@@ -107,28 +118,25 @@ Meteor.publish('commentPost', function(commentId) {
     var comment = Comments.findOne(commentId);
     return Posts.find({_id: comment && comment.post});
   }
+  return [];
 });
 
 // Publish a list of posts
 
-Meteor.publish('postsList', function(find, options) {
+Meteor.publish('postsList', function(terms) {
   if(canViewById(this.userId)){
-    options = options || {};
-    var posts = Posts.find(find, options);
-
+    var parameters = getParameters(terms),
+        posts = Posts.find(parameters.find, parameters.options);
     // console.log('//-------- Subscription Parameters:');
-    // console.log(find);
-    // console.log(options);
+    // console.log(parameters.find);
+    // console.log(parameters.options);
     // console.log('Found '+posts.fetch().length+ ' posts:');
     // posts.rewind();
     // console.log(_.pluck(posts.fetch(), 'headline'));
-    // posts.rewind();
-
     return posts;
   }
+  return [];
 });
-
-
 
 // -------------------------------------------- Comments -------------------------------------------- //
 
@@ -138,6 +146,7 @@ Meteor.publish('postComments', function(postId) {
   if(canViewById(this.userId)){  
     return Comments.find({post: postId});
   }
+  return [];
 });
 
 // Publish a single comment
@@ -146,12 +155,22 @@ Meteor.publish('singleComment', function(commentId) {
   if(canViewById(this.userId)){
     return Comments.find(commentId);
   }
+  return [];
 });
 
 // -------------------------------------------- Other -------------------------------------------- //
 
-Meteor.publish('settings', function() {  
-  return Settings.find();
+Meteor.publish('settings', function() {
+  var options = {};
+  if(!isAdminById(this.userId)){
+    options = _.extend(options, {
+      fields: {
+        mailChimpAPIKey: false,
+        mailChimpListId: false
+      }
+    });
+  }
+  return Settings.find({}, options);
 });
 
 Meteor.publish('notifications', function() {
@@ -159,10 +178,12 @@ Meteor.publish('notifications', function() {
   if(canViewById(this.userId)){
     return Notifications.find({userId:this.userId});
   }
+  return [];
 });
 
 Meteor.publish('categories', function() {
   if(canViewById(this.userId)){
     return Categories.find();
   }
+  return [];
 });