+++ History.md
@@ -1,3 +1,12 @@
+## v0.7.3
+
+* Refactored notifications.
+* Added notifications for new users creation.
+
+## v0.7.2
+
+* Added basic search (thanks Ry!).
+
 ## v0.7.1
 
 * Added karma redistribution.